Create Khan Project 2 Marvel Superheroes
For this project I used advanced SELECT queries to identify data characteristics.
I first selected quantifiable data to find the average height and weights of the superheroes.  In addition I found the tallest and shortest superheroes of the group.
Next I performed a more in focused query by selecting the names, hometowns and heights of superheroes and grouping them by height > 1.8 m.
I also used the CASE WHEN expression to group data by strength conditions and produce a new value. 
Finally, I filtered data by using the AND and OR operators to find the female superheroes who had a neutral alignment and the superheroes who were strong or fast.